requirements.txt
requests
beautifulsoup4
pycryptodome
PySide6
PySide6-Charts
